SSCS-10865 moves panel member exclusions from adjourn case about to submit into issue adjournment notice about to submit
updates tests accordingly
