+++ src/main/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/adjourncase/AdjournCaseAboutToSubmitHandler.java
@@ -2,7 +2,6 @@ package uk.gov.hmcts.reform.sscs.ccd.presubmit.adjourncase;
 
 import static java.util.Objects.nonNull;
 import static uk.gov.hmcts.reform.sscs.ccd.callback.DocumentType.DRAFT_ADJOURNMENT_NOTICE;
-import static uk.gov.hmcts.reform.sscs.ccd.domain.YesNo.YES;
 import static uk.gov.hmcts.reform.sscs.reference.data.model.HearingChannel.PAPER;
 
 import java.time.LocalDate;
@@ -14,13 +13,11 @@ import org.springframework.stereotype.Component;
 import uk.gov.hmcts.reform.sscs.ccd.callback.Callback;
 import uk.gov.hmcts.reform.sscs.ccd.callback.CallbackType;
 import uk.gov.hmcts.reform.sscs.ccd.callback.PreSubmitCallbackResponse;
-import uk.gov.hmcts.reform.sscs.ccd.domain.AdjournCasePanelMembersExcluded;
 import uk.gov.hmcts.reform.sscs.ccd.domain.Adjournment;
 import uk.gov.hmcts.reform.sscs.ccd.domain.DynamicList;
 import uk.gov.hmcts.reform.sscs.ccd.domain.EventType;
 import uk.gov.hmcts.reform.sscs.ccd.domain.Hearing;
 import uk.gov.hmcts.reform.sscs.ccd.domain.HearingOptions;
-import uk.gov.hmcts.reform.sscs.ccd.domain.PanelMemberExclusions;
 import uk.gov.hmcts.reform.sscs.ccd.domain.RegionalProcessingCenter;
 import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseData;
 import uk.gov.hmcts.reform.sscs.ccd.presubmit.PreSubmitCallbackHandler;
@@ -28,7 +25,6 @@ import uk.gov.hmcts.reform.sscs.reference.data.model.HearingChannel;
 import uk.gov.hmcts.reform.sscs.service.AirLookupService;
 import uk.gov.hmcts.reform.sscs.service.PreviewDocumentService;
 import uk.gov.hmcts.reform.sscs.service.RegionalProcessingCenterService;
-import uk.gov.hmcts.reform.sscs.util.SscsUtil;
 
 @Component
 @Slf4j
@@ -104,10 +100,6 @@ public class AdjournCaseAboutToSubmitHandler implements PreSubmitCallbackHandler
 
         updateHearingChannel(sscsCaseData);
 
-        if (isAdjournmentEnabled) {
-            updateExcludedPanelMembers(sscsCaseData);
-        }
-
         return new PreSubmitCallbackResponse<>(sscsCaseData);
     }
 
@@ -135,18 +127,4 @@ public class AdjournCaseAboutToSubmitHandler implements PreSubmitCallbackHandler
                 .findFirst().orElse(HearingChannel.PAPER);
     }
 
-    private static void updateExcludedPanelMembers(SscsCaseData caseData) {
-        Adjournment adjournment = caseData.getAdjournment();
-        AdjournCasePanelMembersExcluded panelMembersExcluded = adjournment.getPanelMembersExcluded();
-        if (nonNull(panelMembersExcluded)) {
-            PanelMemberExclusions panelMemberExclusions = caseData.getSchedulingAndListingFields()
-                .getPanelMemberExclusions();
-
-            if (panelMembersExcluded.equals(AdjournCasePanelMembersExcluded.YES)) {
-                SscsUtil.excludePanelMembers(panelMemberExclusions, adjournment.getPanelMembers());
-            } else if (panelMembersExcluded.equals(AdjournCasePanelMembersExcluded.RESERVED)) {
-                panelMemberExclusions.setArePanelMembersReserved(YES);
-            }
-        }
-    }
 }
